Massive speedup in instrument processing
This takes the 30 + seconds at the beginning to almost zero.

Looks to me like my version does the same thing as the old version based
on the printouts with possibility of decimals on my version.

If agree then switch to my version and startup to image generation is
almost instantaneous.